Validate user cannot cancel reservation if within an hour from start

+++ app/controllers/users/reservations_controller.rb
@@ -4,13 +4,19 @@ class Users::ReservationsController < ApplicationController
 
   def destroy
     reservation = current_user.reservations.find(params[:id])
-    reservation.destroy!
 
-    flash[:success] = "Reservation successfully deleted."
+    if reservation.is_within_one_hour?
+      flash[:success] = "Reservation successfully deleted"
+      reservation.destroy!
+    else
+      flash[:error] = "Reservation starts in less than 1 hour"
+    end
+
     redirect_to user_path
 
   rescue ActiveRecord::RecordNotFound
     flash[:error] = "Reservation not found"
     redirect_to user_path
   end
+
 end
